[td] Use tab from URL
